+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.1.2 / 2017-mm-dd
+==================
+- Relaxed the requirement for pinning to a specific version in constraints. If a
+  requirement is pinned to a version or downloaded from a link, it doesn't need
+  to exist in constraints.
+
 0.1.1 / 2017-02-08
 ==================
 - Raise an exception if requirement and constraint links don't match for a